Added handling for odd number of users
Added logic to handle an odd number of participants and request which
one to give an extra pairing to; also fixed a bug that was causing
pairings to not remove participants from $list2 as they got assigned.

Discovered an unfixed new bug, that if the person selected to get two
partners gets both copies of their name added to $list2, both will be
removed when they are paired the first time causing an unmatched person
at the end.  Setting them as Primary (which will eliminate the
possibility that they will be matched to themselves) will fix this, but
it should be corrected any way.